third_party/WebKit/LayoutTests/editing/deleting/backspace-merge-two-paragraphs.html

<!doctype HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
div {
  border: 1px solid gray;
  padding: 10px;
  line-height: 1.44;
}
</style>
<div contenteditable="true" id="editable">
  <p>This is the first paragraph.</p>
  <p>This is the second.</p>
</div>
<script>
test(function() {
  var editor = document.getElementById('editable');
  var range = document.createRange();
  var selection = window.getSelection();
  range.setStart(editor.childNodes[2], 0);
  range.collapse(true);
  selection.removeAllRanges();
  selection.addRange(range);
  editor.focus();
  document.execCommand('delete', null, false);

  var html_para = document.getElementsByTagName('p')[0].outerHTML;
  assert_equals(html_para, '<p>This is the first paragraph.This is the second.</p>');
}, 'merge two paragraphs by backspace');
</script>
